+++ src/Faker/Providers/email/index.js
@@ -1,0 +1,15 @@
+const extensions = [
+    "@gmail.com",
+    "@hotmail.com",
+    "@hotmail.co.uk",
+    "@hotmail.fr",
+    "@yahoo.com",
+    "@yahoo.co.uk",
+    "@live.com",
+    "@live.co.uk",
+    "@ymail.com",
+    "@outlook.com",
+    "@outlook.co.uk",
+]
+
+module.exports = { extensions }