+++ src/Faker/Providers/title/index.js
@@ -1,0 +1,17 @@
+const maleTitles = [
+    "Mr.",
+    "Master",
+    "Prof.",
+    "Dr.",
+    "Engr.",
+]
+
+const femaleTitles = [
+    "Mrs.",
+    "Ms.",
+    "Prof.",
+    "Dr.",
+    "Engr.",
+]
+
+module.exports = { maleTitles, femaleTitles }