+++ Changelog.md
@@ -1,8 +1,14 @@
 # CHANGELOG
 
+## 2021-10-09, v1.0.2
+- Added test file.
+- Change package code pattern.
+- Added new items to names providers.
+- Email address provider and method added.
+
 ## 2021-10-04, v1.0.1
-- Added more providers for names
-- Code style changes
+- Added more providers for names.
+- Code style changes.
 
 ## 2021-10-03, v1.0.0
-- Initial version
+- Initial version.